Adding changes related to review comments
