+++ src/main/java/org/icgc/argo/program_service/services/ProgramServiceFacade.java
@@ -37,6 +37,7 @@ import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.icgc.argo.program_service.converter.CommonConverter;
+import org.icgc.argo.program_service.converter.DataCenterConverter;
 import org.icgc.argo.program_service.converter.ProgramConverter;
 import org.icgc.argo.program_service.model.dto.DataCenterDTO;
 import org.icgc.argo.program_service.model.entity.JoinProgramInviteEntity;
@@ -57,6 +58,7 @@ public class ProgramServiceFacade {
   private final EgoService egoService;
   private final InvitationService invitationService;
   private final ProgramConverter programConverter;
+  private final DataCenterConverter dataCenterConverter;
   private final CommonConverter commonConverter;
   private final ValidationService validationService;
 
@@ -69,12 +71,14 @@ public class ProgramServiceFacade {
       @NonNull EgoService egoService,
       @NonNull InvitationService invitationService,
       @NonNull ProgramConverter programConverter,
+      @NonNull DataCenterConverter dataCenterConverter,
       @NonNull CommonConverter commonConverter,
       @NonNull ValidationService validationService) {
     this.programService = programService;
     this.egoService = egoService;
     this.invitationService = invitationService;
     this.programConverter = programConverter;
+    this.dataCenterConverter = dataCenterConverter;
     this.commonConverter = commonConverter;
     this.validationService = validationService;
   }
@@ -350,7 +354,7 @@ public class ProgramServiceFacade {
   public List<DataCenterDTO> listDataCenters() {
     val dataCenterEntities = programService.listDataCenters();
     return dataCenterEntities.stream()
-        .map(s -> programConverter.dataCenterToDataCenterEntity(s))
+        .map(s -> dataCenterConverter.dataCenterToDataCenterEntity(s))
         .collect(Collectors.toList());
   }
 }
